requirements.txt
# Core training stack
torch>=2.1.0
transformers>=4.45.0
peft>=0.12.0
trl>=0.12.0
datasets>=2.20.0
bitsandbytes>=0.44.0
accelerate>=0.34.0

# Inference (GGUF — fast CPU)
llama-cpp-python>=0.2.90

# Chatbot UI
gradio>=4.40.0

# Utilities
sentencepiece
protobuf
scipy
huggingface_hub>=0.24.0
safetensors>=0.4.0
